codex: fix CI failure on PR #18632
